Connected input layer on the network svg w/ ellipsis. Huge.
Displays input layer as a part of the network with edges from the input
layer to the first hidden layer. Assigns relative_node_index that
represent new placements for each node. To determine where to place and
are adjusted for ellipsis gaps when layer width is excessive.
Separates large sets into above nodes and below nodes and places
ellipsis in the gap between.
Removed the dummy svgs that were used before.